+++ siriuspy/siriuspy/timesys/time_data/__init__.py
@@ -1,8 +1,6 @@
 """Timesys subpackage."""
 from .connections import IOs, Connections
-from .ll_database import TimingDevDb
 from .hl_types_data import *
 
 del connections
-del ll_database
 del hl_types_data
